refactor: extract isMarketOpen to shared utility function
